+++ submission/submission-webapp/src/itest/kotlin/ac/uk/ebi/biostd/itest/factory/XmlSubmissionFactory.kt
@@ -1,6 +1,5 @@
 package ac.uk.ebi.biostd.itest.factory
 
-import ac.uk.ebi.biostd.itest.assertions.SubmissionSpec
 import ebi.ac.uk.model.Attribute
 import ebi.ac.uk.model.File
 import ebi.ac.uk.model.FilesTable
@@ -10,187 +9,7 @@ import ebi.ac.uk.model.Section
 import ebi.ac.uk.model.SectionsTable
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
-import org.redundent.kotlin.xml.xml
 import org.xmlunit.matchers.EvaluateXPathMatcher.hasXPath
-
-fun submissionSpecXml(accNo: String) = SubmissionSpec(allInOneSubmissionXml(accNo).toString(), fileList().toString())
-
-fun allInOneSubmissionXml(accNo: String) = xml("submission") {
-    attribute("accno", accNo)
-    "attributes" {
-        "attribute" {
-            "name" { -"Title" }
-            "value" { -"venous blood, ∆Monocyte" }
-        }
-        "attribute" {
-            "name" { -"ReleaseDate" }
-            "value" { -"2021-02-12" }
-        }
-    }
-
-    "section" {
-        attribute("accno", "SECT-001")
-        attribute("type", "Study")
-        "attributes" {
-            "attribute" {
-                "name" { -"Project" }
-                "value" { -"CEEHRC (McGill)" }
-            }
-            "attribute" {
-                attribute("reference", true)
-                "name" { -"Organization" }
-                "value" { -"Org1" }
-            }
-            "attribute" {
-                "name" { -"Tissue type" }
-                "value" { -"venous blood" }
-                "valqual" {
-                    "name" { -"Ontology" }
-                    "value" { -"UBERON" }
-                }
-                "nmqual" {
-                    "name" { -"Tissue" }
-                    "value" { -"Blood" }
-                }
-            }
-            "attribute" {
-                "name" { -"File List" }
-                "value" { -"file-list.xml" }
-            }
-        }
-        "links" {
-            "link" {
-                "url" { -"AF069309" }
-                "attributes" {
-                    "attribute" {
-                        "name" { -"type" }
-                        "value" { -"gen" }
-                    }
-                }
-            }
-        }
-        "files" {
-            "file" {
-                "path" { -"DataFile1.txt" }
-                "attributes" {
-                    "attribute" {
-                        "name" { -"Description" }
-                        "value" { -"Data File 1" }
-                    }
-                }
-            }
-            "table" {
-                "file" {
-                    "path" { -"DataFile2.txt" }
-                    "attributes" {
-                        "attribute" {
-                            "name" { -"Description" }
-                            "value" { -"Data File 2" }
-                        }
-                        "attribute" {
-                            "name" { -"Type" }
-                            "value" { -"Data" }
-                        }
-                    }
-                }
-                "file" {
-                    "path" { -"Folder1/DataFile3.txt" }
-                    "attributes" {
-                        "attribute" {
-                            "name" { -"Description" }
-                            "value" { -"Data File 3" }
-                        }
-                        "attribute" {
-                            "name" { -"Type" }
-                            "value" { -"Data" }
-                        }
-                    }
-                }
-                "file" {
-                    "path" { -"Folder1/Folder2/DataFile4.txt" }
-                    "attributes" {
-                        "attribute" {
-                            "name" { -"Description" }
-                            "value" { -"Data File 4" }
-                        }
-                        "attribute" {
-                            "name" { -"Type" }
-                            "value" { -"Data" }
-                        }
-                    }
-                }
-            }
-        }
-        "subsections" {
-            "section" {
-                attribute("accno", "SUBSECT-001")
-                attribute("type", "Stranded Total RNA-Seq")
-                "links" {
-                    "table" {
-                        "link" {
-                            "url" { -"EGAD00001001282" }
-                            "attributes" {
-                                "attribute" {
-                                    "name" { -"Type" }
-                                    "value" { -"EGA" }
-                                }
-                                "attribute" {
-                                    "name" { -"Assay type" }
-                                    "value" { -"RNA-Seq" }
-                                    "valqual" {
-                                        "name" { -"Ontology" }
-                                        "value" { -"EFO" }
-                                    }
-                                    "nmqual" {
-                                        "name" { -"TermId" }
-                                        "value" { -"EFO_0002768" }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            "table" {
-                "section" {
-                    attribute("accno", "DT-1")
-                    attribute("type", "Data")
-                    "attributes" {
-                        "attribute" {
-                            "name" { -"Title" }
-                            "value" { -"Group 1 Transcription Data" }
-                        }
-                        "attribute" {
-                            "name" { -"Description" }
-                            "value" { -"The data for zygotic transcription in mammals group 1" }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-private fun fileList() = xml("table") {
-    "file" {
-        "path" { -"DataFile5.txt" }
-        "attributes" {
-            "attribute" {
-                "name" { -"Type" }
-                "value" { -"referenced" }
-            }
-        }
-    }
-    "file" {
-        "path" { -"Folder1/DataFile6.txt" }
-        "attributes" {
-            "attribute" {
-                "name" { -"Type" }
-                "value" { -"referenced" }
-            }
-        }
-    }
-}
 
 fun assertAllInOneSubmissionXml(xml: String, accNo: String) {
     assertThat(xml, hasXPath("//submission/@accno", equalTo(accNo)))
